feat(aq12): add Orchestra adversarial self-test
