--- .github/workflows/_orchestra-aq12-bootstrap.yml
@@ -1,30 +1,0 @@
-name: Orchestra AQ12 Bootstrap
-
-on:
-  push:
-    branches:
-      - feat/aq12-adversarial-self-test-20260911
-    paths:
-      - 'tools/_aq12_materializer.py'
-      - '.github/workflows/_orchestra-aq12-bootstrap.yml'
-
-permissions:
-  contents: write
-
-jobs:
-  materialize:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout AQ12 branch
-        uses: actions/checkout@v4
-        with:
-          ref: feat/aq12-adversarial-self-test-20260911
-          fetch-depth: 0
-      - name: Set up Python
-        uses: actions/setup-python@v5
-        with:
-          python-version: '3.12'
-      - name: Install targeted validation dependencies
-        run: python -m pip install --upgrade pip pytest "jsonschema[format]>=4.21,<5"
-      - name: Materialize, validate, self-retire, and push AQ12 candidate
-        run: python -B tools/_aq12_materializer.py
