+++ README.json
@@ -2567,6 +2567,12 @@
     "adaptive_assurance_aq11_validator": "scripts/validation/validate_aq11.py",
     "adaptive_assurance_aq11_validation": "tests/runtime/test_adaptive_assurance_aq11.py",
     "adaptive_assurance_aq11_documentation": "docs/architecture/ADAPTIVE_ASSURANCE_AQ11.md",
+    "adaptive_assurance_aq12_contract": "machine/adaptive/aq12-adversarial-self-test.v1.json",
+    "adaptive_assurance_aq12_schema": "machine/schemas/aq12-adversarial-self-test.v1.schema.json",
+    "adaptive_assurance_aq12_runtime": "orchestra_runtime/domain/adaptive/adversarial_self_test.py",
+    "adaptive_assurance_aq12_validator": "scripts/validation/validate_aq12.py",
+    "adaptive_assurance_aq12_validation": "tests/runtime/test_adaptive_assurance_aq12.py",
+    "adaptive_assurance_aq12_documentation": "docs/architecture/ADAPTIVE_ASSURANCE_AQ12.md",
     "comparative_measurement_contract": "machine/benchmarking/comparative-measurement-contract.v1.json",
     "comparative_benchmark_run_schema": "machine/schemas/comparative-benchmark-run.schema.json",
     "comparative_benchmark_experiment_schema": "machine/schemas/comparative-benchmark-experiment.schema.json",
