+++ tests/behavior/run_tests.py
@@ -102,6 +102,7 @@ def main():
         {"Name": "validate_aq9.py", "Path": "scripts/validation/validate_aq9.py"},
         {"Name": "validate_aq10.py", "Path": "scripts/validation/validate_aq10.py"},
         {"Name": "validate_aq11.py", "Path": "scripts/validation/validate_aq11.py"},
+        {"Name": "validate_aq12.py", "Path": "scripts/validation/validate_aq12.py"},
         {"Name": "evaluate_governance.py", "Path": "tests/behavior/evaluate_governance.py"},
         {"Name": "runtime_guardrail.py", "Path": "scripts/runtime_guardrail.py"},
         {"Name": "test_dagger_guardrail.py", "Path": "scripts/test_dagger_guardrail.py"},
